Enabling PRNM code-reuse in triggers!
PRNM's setAllMedPrompts demonstrates how to pass functions from PRNM.
The PRNM function is stringified, and then in the specified function to
be called, the stringified function is eval()'d.

This enables PRNM code-reuse in the actions/trigger-definition code,
enabling e.g. the same logging functions, value-test functions, etc. to
be reused.